A major structural adjustment. Putting all the library .py files to a "ScoreDraft" package directory, with richer Python layer.

+++ UtauDraft/SentenceGenerator_CUDA.cpp
@@ -400,7 +400,6 @@ void SentenceGenerator_CUDA::GenerateSentence(const UtauSourceFetcher& srcFetche
 	float max_srcHalfWinWidth = 0.0f;
 	float max_freqDetectHalfWinWidth = 0.0f;
 
-
 	for (unsigned i = 0; i < numPieces; i++)
 	{
 
@@ -1028,8 +1027,8 @@ void SentenceGenerator_CUDA::GenerateSentence(const UtauSourceFetcher& srcFetche
 			int ipos1 = (int)ceilf(pos_tmpBuf - speed*0.5f);
 			int ipos2 = (int)floorf(pos_tmpBuf + speed*0.5f);
 
-			if (ipos1 >= uTempLen) ipos1 = uTempLen - 1;
-			if (ipos2 >= uTempLen) ipos2 = uTempLen - 1;
+			if (ipos1 >= (int)uTempLen) ipos1 = (int)uTempLen - 1;
+			if (ipos2 >= (int)uTempLen) ipos2 = (int)uTempLen - 1;
 
 			float sum = 0.0f;
 			for (int ipos = ipos1; ipos <= ipos2; ipos++)
